29/02/2024 primera

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Hash;
 
 class UsersController extends Controller
 {
@@ -15,20 +16,42 @@ class UsersController extends Controller
     public function show($id){
         return User::find($id);
     }
+    //Crear usuario
 
     public function store(Request $request){
-        return User::create($request ->all());
+        $this->validate($request, [
+            'username' => 'required|unique:users',
 
-    }
-    public function update(Request $request, $id){
-        $user = User::find($id);
+            'password' => 'required'
+        ]);
+        $user = new User;
         $user->fill($request->all());
+        $user->password = Hash::make($request->password);
         $user->save();
         return $user;
+
     }
+   //actualizar un usuario
+   public function update(Request $request, $id){
+    //if($request->username){
+   //    $exist = User::where('username',$request->username);
+    //       if($exist) return response()->json(['Error el dato ya existe'], 400);
+  //  }
+    $this->validate($request,[
+       'username' => 'required|unique:users,username,'.$id,
+   ]);
+    $user = User::find($id);
+    if(!$user) return response('', 404);
+    $user->fill($request->all());
+    if($request->password)
+          $user->password = Hash::make($request->password);
+    $user->save();
+    return $user;
+}
 
     public function destroy( $id){
         $user = User::find($id);
+        if(!$user) return response('', 404);
         $user->delete();
         return $user;
     }
